@@ -1,28 +1,28 @@
-const admin = require("firebase-admin");
-const functions = require("firebase-functions");
-const axios = require("axios");
-const jwt = require("jsonwebtoken");
-const jwksClient = require("jwks-rsa");
-const { GLOBAL_CONFIG } = require("../global-config-functions");
+import { getAuth } from "firebase-admin/auth";
+import axios from "axios";
+import { JwksClient } from "jwks-rsa";
+import { decode, verify } from "jsonwebtoken";
+import { error, log } from "firebase-functions/logger";
+import { onCall } from "firebase-functions/v1/https";
 
-const { FIREBASE_SERVER_REGION } = GLOBAL_CONFIG
+import { FIREBASE_SERVER_REGION } from "../global-config-functions";
 
-const client = jwksClient({
+const client = JwksClient({
   jwksUri: "https://login.eveonline.com/oauth/jwks",
 });
 
-exports.updateCorpIDs = functions
-  .region(FIREBASE_SERVER_REGION)
-  .https.onCall(async (data, context) => {
+const updateCorpIDs = onCall(
+  { region: FIREBASE_SERVER_REGION },
+  async (data, context) => {
     async function setClaim() {
       let corpIDs = new Set();
 
       for (let charData of data) {
-        const decodedToken = jwt.decode(charData.authToken);
+        const decodedToken = decode(charData.authToken);
         const kid = decodedToken.kid;
         const key = await client.getSigningKey(kid);
         const getSigningKey = key.getPublicKey();
-        let tokenState = jwt.verify(
+        let tokenState = verify(
           charData.authToken,
           getSigningKey,
           (err, decoded) => {
@@ -30,10 +30,8 @@ exports.updateCorpIDs = functions
               decoded.iss != "login.eveonline.com" &&
               decoded.iss != "https://login.eveonline.com"
             ) {
-              functions.logger.error(
-                `${context.auth.uid} failed to verify Eve Token`
-              );
-              functions.logger.error(JSON.stringify(decoded));
+              error(`${context.auth.uid} failed to verify Eve Token`);
+              error(JSON.stringify(decoded));
               return false;
             } else {
               return true;
@@ -52,17 +50,20 @@ exports.updateCorpIDs = functions
           }
         }
       }
-      functions.logger.log(`${context.auth.uid} Corporation Claims Updated`);
-      await admin.auth().setCustomUserClaims(context.auth.uid, {
+      log(`${context.auth.uid} Corporation Claims Updated`);
+      await getAuth().setCustomUserClaims(context.auth.uid, {
         corporations: [...corpIDs],
       });
       return;
     }
 
     if (!context.auth) {
-      functions.logger.error("Unathorised Claims User");
+      error("Unathorised Claims User");
       return { error: "Unauthorised User" };
     }
     await setClaim();
     return null;
-  });
+  }
+);
+
+export default updateCorpIDs;
