@@ -10,6 +10,12 @@ import {
 import { useJobManagement } from "../../../../../../Hooks/useJobManagement";
 import { useMarketOrderFunctions } from "../../../../../../Hooks/GeneralHooks/useMarketOrderFunctions";
 import { useHelperFunction } from "../../../../../../Hooks/GeneralHooks/useHelperFunctions";
+import {
+  LARGE_TEXT_FORMAT,
+  STANDARD_TEXT_FORMAT,
+  TWO_DECIMAL_PLACES,
+  ZERO_DECIMAL_PLACES,
+} from "../../../../../../Context/defaultValues";
 
 class ESIMarketOrder {
   constructor(order) {
@@ -72,7 +78,10 @@ export function AvailableMarketOrdersTab({
             const charData = users.find(
               (i) => i.CharacterHash === order.CharacterHash
             );
-            const locationData = findUniverseItemObject(order.location_id);
+            const locationName =
+              findUniverseItemObject(order.location_id)?.name ||
+              "Location Data Unavailable";
+
             const corpData = corpEsiData.get(charData?.corporation_id);
 
             return (
@@ -104,10 +113,10 @@ export function AvailableMarketOrdersTab({
                       <Avatar
                         src={
                           order.is_corporation
-                            ? corpData !== undefined
+                            ? corpData
                               ? `https://images.evetech.net/corporations/${corpData.corporation_id}/logo`
                               : ""
-                            : charData !== undefined
+                            : charData
                             ? `https://images.evetech.net/characters/${charData.CharacterID}/portrait`
                             : ""
                         }
@@ -120,54 +129,42 @@ export function AvailableMarketOrdersTab({
                     </Tooltip>
                     <Grid item xs={12}>
                       <Typography variant="body2">
-                        {order.volume_remain.toLocaleString(undefined, {
-                          minimumFractionDigits: 0,
-                          maximumFractionDigits: 0,
-                        })}
+                        {order.volume_remain.toLocaleString(
+                          undefined,
+                          ZERO_DECIMAL_PLACES
+                        )}
                         /
-                        {order.volume_total.toLocaleString(undefined, {
-                          minimumFractionDigits: 0,
-                          maximumFractionDigits: 0,
-                        })}{" "}
+                        {order.volume_total.toLocaleString(
+                          undefined,
+                          ZERO_DECIMAL_PLACES
+                        )}{" "}
                         Items Remaining
                       </Typography>
                     </Grid>
                     <Grid item xs={12}>
-                      <Typography
-                        sx={{ typography: { xs: "caption", sm: "body2" } }}
-                      >
-                        {order.price.toLocaleString(undefined, {
-                          minimumFractionDigits: 2,
-                          maximumFractionDigits: 2,
-                        })}{" "}
+                      <Typography sx={{ typography: STANDARD_TEXT_FORMAT }}>
+                        {order.price.toLocaleString(
+                          undefined,
+                          TWO_DECIMAL_PLACES
+                        )}{" "}
                         ISK Per Item
                       </Typography>
                     </Grid>
                     <Grid item xs={12}>
-                      <Typography
-                        sx={{ typography: { xs: "caption", sm: "body2" } }}
-                      >
-                        {locationData !== undefined
-                          ? locationData.name
-                          : "Location Data Unavailable"}
+                      <Typography sx={{ typography: STANDARD_TEXT_FORMAT }}>
+                        {locationName}
                       </Typography>
                     </Grid>
                     <Grid item xs={12}>
-                      <Typography
-                        sx={{ typography: { xs: "caption", sm: "body2" } }}
-                      >
+                      <Typography sx={{ typography: STANDARD_TEXT_FORMAT }}>
                         Duration: {order.duration} Days
                       </Typography>
                     </Grid>
                     <Grid item xs={12} sx={{ marginTop: "5px" }}>
-                      <Typography
-                        sx={{ typography: { xs: "caption", sm: "body2" } }}
-                      >
+                      <Typography sx={{ typography: STANDARD_TEXT_FORMAT }}>
                         Last Modified:
                       </Typography>
-                      <Typography
-                        sx={{ typography: { xs: "caption", sm: "body2" } }}
-                      >
+                      <Typography sx={{ typography: STANDARD_TEXT_FORMAT }}>
                         {new Date(order.issued).toLocaleString()}
                       </Typography>
                     </Grid>
@@ -259,7 +256,7 @@ export function AvailableMarketOrdersTab({
           })
         ) : (
           <Grid item xs={12} align="center">
-            <Typography sx={{ typography: { xs: "body2", md: "body1" } }}>
+            <Typography sx={{ typography: LARGE_TEXT_FORMAT }}>
               There are no orders appearing on the API matching this item type.
             </Typography>
           </Grid>
