@@ -7,12 +7,14 @@ import {
   Typography,
 } from "@mui/material";
 import { useContext } from "react";
-import { UsersContext } from "../../../../Context/AuthContext";
 import { useFirebase } from "../../../../Hooks/useFirebase";
+import { ApplicationSettingsContext } from "../../../../Context/LayoutContext";
 
 export function CompactLayoutSettings({ parentUserIndex }) {
-  const { users, updateUsers } = useContext(UsersContext);
-  const { updateMainUserDoc } = useFirebase();
+  const { applicationSettings, updateApplicationSettings } = useContext(
+    ApplicationSettingsContext
+  );
+  const { uploadApplicationSettings } = useFirebase();
 
   return (
     <Paper elevation={3} sx={{ padding: "20px" }} square={true}>
@@ -29,16 +31,14 @@ export function CompactLayoutSettings({ parentUserIndex }) {
                 control={
                   <Switch
                     checked={
-                      !users[parentUserIndex].settings.layout.hideTutorials
+                      !applicationSettings.hideTutorials
                     }
                     color="primary"
                     onChange={(e) => {
-                      let newUsersArray = [...users];
-                      newUsersArray[
-                        parentUserIndex
-                      ].settings.layout.hideTutorials = !e.target.checked;
-                      updateUsers(newUsersArray);
-                      updateMainUserDoc();
+                      const newApplicationSettings =
+                        applicationSettings.toggleHideTutorials();
+                      updateApplicationSettings(newApplicationSettings);
+                      uploadApplicationSettings(newApplicationSettings);
                     }}
                   />
                 }
@@ -55,17 +55,14 @@ export function CompactLayoutSettings({ parentUserIndex }) {
                 control={
                   <Switch
                     checked={
-                      users[parentUserIndex].settings.layout?.enableCompactView
+                      applicationSettings.enableCompactView
                     }
                     color="primary"
                     onChange={(e) => {
-                      let newUsersArray = [...users];
-                      newUsersArray[
-                        parentUserIndex
-                      ].settings.layout.enableCompactView = e.target.checked;
-
-                      updateUsers(newUsersArray);
-                      updateMainUserDoc();
+                      const newApplicationSettings =
+                        applicationSettings.toggleEnableCompactView();
+                      updateApplicationSettings(newApplicationSettings);
+                      uploadApplicationSettings(newApplicationSettings);
                     }}
                   />
                 }
